+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [Added] scrape bt-agenda
 * [Added] Add support for scraping currentState history
+* [Search] Use elastic-search server [#248](https://github.com/demokratie-live/democracy-client/issues/248)
 
 ### 0.1.0